backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-simple-history==3.4.0
djangorestframework-simplejwt==5.3.1
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
python-dotenv==1.0.0
boto3==1.28.0